test(browser): isolate foreground lease hooks from production
