+++ packages/jazz-tools/tests/browser/jazz-broker-worker-test.ts
@@ -1,0 +1,32 @@
+import { installJazzBrokerWorker } from "../../src/worker/jazz-broker-worker-core.js";
+import type {
+  BrowserForegroundNodeLeaseAcquireRequest,
+  BrowserForegroundNodeLeaseAcquireResponse,
+} from "../../src/runtime/native-runtime/browser-worker-protocol.js";
+
+// This entry is served only by the browser test harness. Production bundles
+// start `src/worker/jazz-broker-worker.ts`, which installs no hooks and never
+// inspects these test-only request fields.
+installJazzBrokerWorker({
+  foregroundLeaseTestHooks: {
+    delayAfterLeaseAllocation(
+      request: BrowserForegroundNodeLeaseAcquireRequest,
+    ): number | undefined {
+      const delay = request.testDelayAfterLeaseAllocationMs;
+      if (delay === undefined) return undefined;
+      if (!Number.isSafeInteger(delay) || delay < 0 || delay > 1_000) {
+        throw new Error("Invalid foreground lease test delay");
+      }
+      return delay;
+    },
+    allocationCommitted(port, node) {
+      port.postMessage({
+        type: "foreground-node-lease-test-allocated",
+        node,
+      } satisfies BrowserForegroundNodeLeaseAcquireResponse);
+    },
+    async cancellationRetired(pageStore, node) {
+      return await pageStore.foregroundNodeLeaseNodeState(node);
+    },
+  },
+});
